feat: replace drawer with task modal and add history
