+++ taintedpaint/app/api/jobs/route.ts
@@ -101,15 +101,16 @@ export async function POST(req: NextRequest) {
     });
     await Promise.all(pendingWrites);
 
-    const {
-      customerName = '',
-      representative = '',
-      inquiryDate = '',
-      deliveryDate = '',
-      ynmxId = '',
-      notes = '',
-      folderName = '',
-    } = fields;
+      const {
+        customerName = '',
+        representative = '',
+        inquiryDate = '',
+        deliveryDate = '',
+        ynmxId = '',
+        notes = '',
+        folderName = '',
+        userName = '',
+      } = fields;
 
     if (
       tempFiles.length === 0 ||
@@ -161,6 +162,15 @@ export async function POST(req: NextRequest) {
       deliveryNoteGenerated: false,
       awaitingAcceptance: false,
       updatedAt: new Date().toISOString(),
+      history: userName
+        ? [
+            {
+              user: userName,
+              action: '创建任务',
+              timestamp: new Date().toISOString(),
+            },
+          ]
+        : [],
     };
 
     await updateBoardData(async (boardData) => {
